working on poll() and have it working for 1 connection with multiple communications
