--- default.conf
@@ -1,43 +1,0 @@
-# Default server configuration
-server {
-	listen 8081;
-	host	localhost;
-	server_name default_server;
-	error_page 404 /errors/404.html;
-	error_page 500 /error/500.html;
-	client_max_body_size 1000000;
-
-	location / {
-		root /var/www/html;
-		index index.html;
-		autoindex on;
-		allowed_methods GET POST DELETE;
-	}
-
-	# 	Redirect example
-	location /old {
-#		redirect /new;
-		redirect http://example.com/new;
-	}
-
-# 	Static files location
-	location /static {
-		root /var/www/static;
-		allowed_methods GET;
-		autoindex off;
-	}
-
-	location /upload {
-		root /var/www/uploads;
-		upload_path /var/www/uploads;
-		allowed_methods POST;
-	}
-
-	location /cgi-bin {
-		root /var/www/cgi;
-		cgi_path /usr/bin/python3;
-		allowed_methods GET POST;
-		cgi .php /usr/bin/php-cgi;
-		cgi .py /usr/bin/python3;
-	}
-}
